refactor(file-upload): rename sgds-fileupload to sgds-file-upload
